+++ packages/core/blanket/examples/01-click-through.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import Button from '@atlaskit/button';
 import Blanket from '../src';