+++ packages/core/button/examples/10-Button.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import Button from '../src';
 
 export default () => <Button>Button</Button>;