+++ packages/core/section-message/src/components/SectionMessage.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import Button from '@atlaskit/button';
 import { baseAppearanceObj } from '../theme';
 import { Appearance } from '../types';