+++ packages/elements/mention/src/components/MentionSpotlight/index.tsx
@@ -1,22 +1,25 @@
-import * as React from 'react';
-import * as Styled from './styles';
+import React from 'react';
 
 import Button from '@atlaskit/button';
 import EditorCloseIcon from '@atlaskit/icon/glyph/editor/close';
-
 import Tooltip from '@atlaskit/tooltip';
+
 import {
   SpotlightTitle,
   SpotlightCloseTooltip,
   SpotlightDescription,
   SpotlightDescriptionLink,
 } from '../../util/i18n';
+import * as Styled from './styles';
 
 export interface Props {
   createTeamLink: string;
   /** Callback to track the event where user click on x icon */
   onClose: () => void;
 }
+
+const ICON_URL =
+  'https://ptc-directory-sited-static.us-east-1.prod.public.atl-paas.net/teams/avatars/2.svg';
 
 export default class MentionSpotlight extends React.Component<Props, {}> {
   render() {
@@ -26,7 +29,7 @@ export default class MentionSpotlight extends React.Component<Props, {}> {
       <Styled.Card>
         <Styled.Content>
           <Styled.Aside>
-            <img src={Styled.iconUrl} height={32} />
+            <img src={ICON_URL} height={32} />
           </Styled.Aside>
           <Styled.Section>
             <Styled.Heading>
@@ -43,7 +46,7 @@ export default class MentionSpotlight extends React.Component<Props, {}> {
                       {linkText => (
                         <a href={createTeamLink} target="_blank">
                           {' '}
-                          {linkText}{' '}
+                          {linkText}
                         </a>
                       )}
                     </SpotlightDescriptionLink>