+++ packages/core/datetime-picker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/datetime-picker
 
+## 8.0.9
+
+- Updated dependencies [790e66bece](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/790e66bece):
+  - @atlaskit/button@13.0.11
+  - @atlaskit/modal-dialog@10.0.10
+  - @atlaskit/select@10.0.0
+
 ## 8.0.8
 
 ### Patch Changes