+++ packages/core/inline-message/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/inline-message
 
+## 10.0.8
+
+### Patch Changes
+
+- [patch][29a1f158c1](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/29a1f158c1):
+
+  Use default react import in typescript files.
+
 ## 10.0.7
 
 - Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):