+++ packages/core/logo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/logo
 
+## 12.1.1
+
+- Updated dependencies [790e66bece](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/790e66bece):
+  - @atlaskit/button@13.0.11
+  - @atlaskit/select@10.0.0
+
 ## 12.1.0
 
 ### Minor Changes